refactor(feature): move direct_method_stream and desired_properties_stream to feature/command
Both are stateless stream constructors that transform cloud receivers
into CommandRequestStream values. They belong alongside interval_stream
and debounced_command_stream rather than as Twin methods.
